http: `CurlCookie` as a separate file
based on the review
